# test/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/recipe": {
      "post": {
        "operationId": "RecipeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRecipeDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      },
      "get": {
        "operationId": "RecipeController_findAll",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      }
    },
    "/recipe/{id}": {
      "get": {
        "operationId": "RecipeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      },
      "patch": {
        "operationId": "RecipeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRecipeDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } }
      },
      "delete": {
        "operationId": "RecipeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    },
...
  },
  "info": {
    "title": "nestJs",
    "description": "The nestJs API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "", "description": "dev" },
    { "url": "http://localhost:3001", "description": "local" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreateRecipeDto": { "type": "object", "properties": {} },
      "UpdateRecipeDto": { "type": "object", "properties": {} },
...
    }
  },
  "security": [{ "bearer": [] }]
}